IOSProtoFileWriter.kt now uses multiplatform coded input/output stream and calculations.

+++ grpc-multiplatform-lib/src/androidJvmCommon/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/KMMessage.kt
@@ -5,7 +5,7 @@ import java.nio.ByteBuffer
 
 actual interface KMMessage {
 
-    val requiredSize: Long
+    val requiredSize: Int
 
     fun serialize(): ByteArray {
         val data = ByteArray(requiredSize.toInt())
